Fix neighbour searching when max_refinement_level > 0

+++ tests/game_of_life/scalability.cpp
@@ -128,7 +128,7 @@ int main(int argc, char* argv[])
 		}
 	}
 	after = clock();
-	total += double(after) - before;
+	total += after - before;
 	if (comm.rank() == 0) {
 		cout << endl;
 	}
